Clearer microstepping for quickly rearranging packings
Some packings rearrange for very small strains - smaller than the initially
attempted strain. For these, we go to a much smaller strain (ca. 1e-16 instead
of 1e-8 or so), so that even these strains are determined correctly.

This code was already implemented, but not in a very clear way. This commit
improves this.